Divide root file for atom and slector exercise
