+++ src/components/WordCounter/CharactherCounter.js
@@ -1,29 +1,8 @@
-import { useRecoilState } from "recoil"
-
-const textState = atom({
-  key: 'textState',
-  default: '',
-})
+ import { TextInput } from './TextInput';
 
 export const characterCounter = () => {
   return (
     <TextInput /> 
-    <characterCounter />
   )
 }
 
-export const textInput = () => {
-  const [text, setText] = useRecoilState(textState);
-
-  const onChange = (event) => {
-    setText(event.target.value);
-  };
-
-  return (
-    <div>
-      <input type='text' value={text} onChange={onChange} />
-      <br />
-      Echo: {text}
-    </div>
-  )
-}
